requirements.txt
websocket-client>=1.6.0

